Kurakin Matvey. FIIT1. Lab 1. Clang AST (#10)
The task was to analyze user-defined classes and structures
(user-defined types)

For user-defined types, the base types are defined

The type and access modifier are defined for fields.

The methods are defined by the return type and the types of parameters
they accept, the access modifier, virtual, pure virtual, or override.

Implicitly created methods are ignored