Lopatin Ilya. Lab 4: MLIR. Option 3. (#109)
This is a transformation pass designed for the MLIR framework. Its
primary purpose is to lower arithmetic remainder operations,
specifically `arith.remsi` and `arith.remui`, into a sequence of more
fundamental operations.

The pass replaces each `arith.remsi` or `arith.remui` operation with an
equivalent expression based on the mathematical identity:
```
rem(a, b) = a - (a // b) * b
```
Here, `//` denotes integer division, which corresponds to `arith.divsi`
for signed integers and `arith.divui` for unsigned integers.

The pass is implemented as an `OpRewritePattern` and operates at the
module level. It automatically identifies and transforms applicable
operations within the module.

--- 
I'm sorry to tell you about my life circumstances, but I will not be
able to answer you on saturdays anymore, nor with this lab or the third
one :(

---------

Co-authored-by: Kuznetsov-Artyom <110616662+Kuznetsov-Artyom@users.noreply.github.com>